+++ packages/atomic/src/components/facets/atomic-numeric-facet/atomic-numeric-facet.tsx
@@ -54,6 +54,7 @@ export class AtomicNumericFacet
   public strings: I18nState = {
     clear: () => this.bindings.i18n.t('clear'),
     facetValue: (variables) => this.bindings.i18n.t('facetValue', variables),
+    to: (variables) => this.bindings.i18n.t('to', variables),
   };
 
   @State() public isExpanded = false;
@@ -66,12 +67,8 @@ export class AtomicNumericFacet
    * The non-localized label for the facet.
    */
   @Prop() public label = 'No label';
-  /**
-   * Whether or not the index should automatically generate options for the facet.
-   */
-  @Prop() public generateAutomaticRanges = true;
 
-  public buildOptions() {
+  private buildManualRanges() {
     const options = Array.from(
       this.host.querySelectorAll('atomic-numeric-range')
     );
@@ -81,17 +78,20 @@ export class AtomicNumericFacet
   }
 
   public initialize() {
+    const manualRanges = this.buildManualRanges();
     const options: NumericFacetOptions = {
       facetId: this.facetId,
       field: this.field,
-      generateAutomaticRanges: this.generateAutomaticRanges,
-      currentValues: this.buildOptions(),
+      generateAutomaticRanges: manualRanges.length === 0,
+      currentValues: manualRanges,
     };
 
     this.facet = buildNumericFacet(this.bindings.engine, {options});
     this.strings[this.label] = () => this.bindings.i18n.t(this.label);
     this.facetId = this.facet.state.facetId;
-    this.bindings.store.state.facetLabels[this.facetId] = this.label;
+    this.bindings.store.state.facets[this.facetId] = {
+      label: this.label,
+    };
   }
 
   private get values() {
@@ -105,15 +105,23 @@ export class AtomicNumericFacet
 
   private buildListItem(item: NumericFacetValue) {
     const isSelected = this.facet.isValueSelected(item);
+    const {language} = this.bindings.i18n;
+    const start = item.start.toLocaleString(language);
+    const end = item.end.toLocaleString(language);
+    const value = this.strings.to({start, end});
+
     return (
       <FacetValue
-        label={` ${item.start}-${item.end}`}
+        label={value}
         isSelected={isSelected}
         numberOfResults={item.numberOfResults}
         facetValueSelected={() => {
           this.facet.toggleSelect(item);
         }}
-        ariaLabel={this.strings.facetValue(item)}
+        ariaLabel={this.strings.facetValue({
+          value,
+          numberOfResults: item.numberOfResults,
+        })}
       />
     );
   }