+++ packages/headless/src/api/search/search-api-client.ts
@@ -62,7 +62,7 @@ export class SearchAPIClient {
   ): Promise<SearchAPIClientResponse<PlanResponseSuccess>> {
     const platformResponse = await PlatformClient.call<PlanResponseSuccess>({
       ...baseSearchRequest(req, 'POST', 'application/json', '/plan'),
-      requestParams: pickNonBaseParams(req) as PlanRequest, // TODO: This cast won't be needed once all methods have been reworked and we can change types in PlatformClient
+      requestParams: pickNonBaseParams(req),
       ...this.options,
     });
 
@@ -211,31 +211,32 @@ export class SearchAPIClient {
   }
 }
 
-const unwrapError = (res: PlatformResponse<AllSearchAPIResponse>) => {
-  if (isException(res)) {
-    return unwrapByBodyException(res);
+const unwrapError = (
+  res: PlatformResponse<AllSearchAPIResponse>
+): SearchAPIErrorWithStatusCode => {
+  if (isSearchAPIException(res)) {
+    return unwrapErrorByException(res);
   }
-  if (isError(res)) {
-    return unwrapByStatusCode(res);
+  if (isSearchAPIErrorWithStatusCode(res)) {
+    return res.body;
   }
 
-  return {message: 'unknown', statusCode: 0, type: 'unknown'};
+  const body = (res.body as unknown) as Error;
+
+  return {
+    ...body,
+    message: `Client side error: ${body.message || ''}`,
+    statusCode: 400,
+    type: 'ClientError',
+  };
 };
 
-const unwrapByBodyException = (
+const unwrapErrorByException = (
   res: PlatformResponse<SearchAPIErrorWithExceptionInBody>
-) => ({
+): SearchAPIErrorWithStatusCode => ({
   message: res.body.exception.code,
   statusCode: res.response.status,
   type: res.body.exception.code,
-});
-
-const unwrapByStatusCode = (
-  res: PlatformResponse<SearchAPIErrorWithStatusCode>
-) => ({
-  message: res.body.message,
-  statusCode: res.body.statusCode,
-  type: res.body.type,
 });
 
 export const isSuccessResponse = <T>(
@@ -282,7 +283,7 @@ function isSuccessSearchResponse(
   );
 }
 
-function isError(
+function isSearchAPIErrorWithStatusCode(
   r: PlatformResponse<AllSearchAPIResponse>
 ): r is PlatformResponse<SearchAPIErrorWithStatusCode> {
   return (
@@ -291,7 +292,7 @@ function isError(
   );
 }
 
-function isException(
+function isSearchAPIException(
   r: PlatformResponse<AllSearchAPIResponse>
 ): r is PlatformResponse<SearchAPIErrorWithExceptionInBody> {
   return (